+++ WorkSynergy.Core.Application/Features/Tags/Commands/DeleteTag/DeleteTagCommand.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
 using MediatR;
+using Microsoft.AspNetCore.Http;
+using WorkSynergy.Core.Application.Exceptions;
 using WorkSynergy.Core.Application.Interfaces.Repositories;
 using WorkSynergy.Core.Application.Wrappers;
 
@@ -23,9 +25,14 @@ namespace WorkSynergy.Core.Application.Features.Tags.Commands.DeleteTag
         public async Task<Response<int>> Handle(DeleteTagCommand request, CancellationToken cancellationToken)
         {
             Response<int> response = new();
-            response.Succeeded = true;
             var post = await _tagRepository.GetByIdAsync(request.Id);
+            if (post == null)
+            {
+                throw new ApiException("Post not found", StatusCodes.Status404NotFound);
+            }
             await _tagRepository.DeleteAsync(post);
+            response.Succeeded = true;
+            response.StatusCode = StatusCodes.Status204NoContent;
             return response;
         }
     }
